Removed the .o4d container format and consolidated requirements
The .o4d container, its converters, and its players are gone. open4d/io,
open4d/tools, and open4d/player existed only to read, write, or play .o4d,
so they are deleted outright along with the three examples/play_*_o4d.py
scripts and tests/test_o4d_mesh_sequence.py.

The Open3D integration is trimmed rather than removed. frame_to_open3d
converts any frame-like object and is not .o4d-specific, so it stays; the
load_frame/iter_frames/sequence_info loaders and the container header structs
are gone, 270 lines down to 156. Its four generic tests are kept and the three
.o4d-backed ones removed. open4d/core was untouched — it never imported io.

DracoPy was used only by the .o4d Draco point-cloud reader, so the draco extra
is dead and removed from it and from all. codecs/draco is unaffected: it uses
the Draco C++ binaries, not DracoPy.

Requirements now live in one place. The root README gained a Requirements
section with a single baseline for the repo plus a table of what each codec and
the RGB-D pipeline add on top, since they conflict with one another on Python
version alone and Open3D's missing 3.13 wheels cap every one of them at 3.12.
The RGB-D README lost its Requirements and Tested environment sections (~70
lines of host specs and measured operating points) and now covers only how to
run the pipeline; its C++ dependency list moved into the root table rather than
being dropped. The calibration walkthrough was nested inside the deleted
section and was preserved.

The memory figure in the baseline table is measured, not estimated: 1.06 MB per
frame for a 20,672-vertex / 39,421-triangle mesh. Measuring it surfaced that
shade() returned float64 because a Python tuple of floats promoted the largest
per-vertex buffer; forcing float32 halves the colour array, 6.65 MB to 3.33 MB
across ten frames, with no visible change.

Also fixed: the repository-layout diagram omitted apps/ and basketball_app/;
examples/visualization/README.md documented a --yaw flag that does not exist,
so that command failed outright; the root README still described the replaced
Open3D viewer and linked to a nonexistent examples/README.md. basketball_app
was removed on request, and requires-python = ">=3.10" now enforces the
version floor the docs claim. Two committed .o4d .pyc files and the two
generated egg-info directories are untracked, with *.egg-info/ ignored.

Verified: 12/12 core tests pass, the viewer plays and saves an identical GIF,
and no tracked text file mentions .o4d. The four Open3D adapter tests were not
run — Open3D has no 3.13 wheel, which is the ceiling documented above.

+++ examples/visualization/render_frames.py
@@ -127,8 +127,13 @@ def shade(
     # An ambient floor keeps faces turned away from the light off black.
     intensity = ambient + (1.0 - ambient) * diffuse
 
-    shaded = np.clip(intensity[:, None] * np.asarray(base), 0.0, 1.0)
-    return np.column_stack([shaded, np.ones(len(positions), dtype=np.float32)])
+    # float32 throughout: a Python tuple of floats would promote this to float64
+    # and double the largest per-vertex buffer for no visible gain.
+    shaded = np.clip(intensity[:, None] * np.asarray(base, dtype=np.float32),
+                     0.0, 1.0)
+    return np.column_stack(
+        [shaded, np.ones(len(positions), dtype=np.float32)]
+    ).astype(np.float32)
 
 
 def vertex_colors(
